Allow (xc)frameworks with unusual binary locations
Summary:
Most frameworks adopt the convention of having their binaries named like so:

`MyFramework.framework/MyFramework`

However, there's no obligation to do this, and a different location can be used by specifying it in the Info.plist for the framework, or a top-level Info.plist for an Xcframework.

In the case of one such framework (D83315376) the binary paths are only specified in the Xcframework Info.plist.

The way `prebuilt_apple_xcframework` works is to find the path to the correct framework for the current target platform, and return a `prebuilt_apple_framework` for that path. This inherently loses the information in the Info.plist for the Xcframework, so it can't be used at build time to find the binary.

The simplest solution is to specify the binary path inside the rule. If it varies by platform, we can use a Select. Seems sensible to call this `binary`?

Reviewed By: rmaz

Differential Revision: D83328100

fbshipit-source-id: 11732e614c8c6081ab4a19f34531b30c7b1561ce