explicitly return all input source files for generated targets
Summary:
## Context

We now do automated code gen for e.g. thrift files.

Because we run `elp.bxl` in a buck2 isolation directory, the output that we use is not updated by a normal buck2 build, since our file watchers are set on the isolation directory outputs.

The codegen itself is re-run when any of the inputs for a generated target change, and we do not currently track these in the ELP target model.

## This diff

Update `elp.bxl` so that for targets which include the "generated" label we return a field containing the transitive set of inputs used in the target.

Reviewed By: TheGeorge, robertoaloi

Differential Revision: D83155666

fbshipit-source-id: f0192c73e2db290f2cb0c68fe0ddf8303354097b