Enable content-based path hashing for CUDA compile actions
Summary:
Mostly straightforward plumbing of the `content_based_hashing` flag, except one issue in the `cuda_cxx_compile` action.

As an intermediate lowering step, `cudafepp` is an action that generates a cpp file `.cudafe1.cpp` which includes a couple stub files (`.cudafe1.stub.c` and `.fatbin.c`). These includes are hardcoded in the cpp file and its dependencies in the form of file basenames like this:

```
#include "add.cu.o.compute_90a.cudafe1.stub.c"
```

The stubs are generated by other actions from the dist NVCC actions. Ideally these includes shouldn't be hardcoded but should use the `OutputArtifact`s of the producing actions instead. However, those actions depend on `cudafepp` so doing so would result in a circular dependency.

This pattern works before content-based hashing because everything is put under the same directory. It doesn't work with content-based hashing now because the outputs are put under the `output_artifacts` sub-directory so the includes cannot be found.

To work around this, we create a directory containing symlinks to the actual stub files using `symlinked_dir`. And then we add the directory to the C++ compiler search paths using `-I`.

Reviewed By: get9

Differential Revision: D82593373

fbshipit-source-id: c8902e858ff86b07eb6f6d10d0f815a0f28f166f