Enable tranformation spec for rust with shared linking
Summary:
Transformation specs should also be available for Rust (mostly to migrate off link group-based APIs that is currently rolled out for rust too). Thankfully this is easy to configure because all we should do is to propagate the values and all handling will work out of the box.

Passing transformation spec context to part that handles shared linkage

Reviewed By: dtolnay

Differential Revision: D83559460

fbshipit-source-id: 8484d265c96f10ae721431f26fec18a49e01ec41