Enable transformation spec for rust with link groups
Summary:
Transformation specs should also be available for Rust (mostly to migrate off link group-based APIs that is currently rolled out for rust too). Thankfully this is easy to configure because all we should do is to propagate the values and all handling will work out of the box.

Passing transformation spec context to part that handles link groups

Reviewed By: dtolnay

Differential Revision: D83559462

fbshipit-source-id: f6095b7e81bdc1ebec58b77f7b78643bdcad2eee