Allow transformation spec to be passed to rust binaries
Summary:
Transformation specs should also be available for Rust (mostly to migrate off link group-based APIs that is currently rolled out for rust too). Thankfully this is easy to configure because all we should do is to propagate the values and all handling will work out of the box.

Allowing transformation spec to be configured for rust binaries and tests

Reviewed By: yahayaohinoyi

Differential Revision: D83559454

fbshipit-source-id: a5e27f1d0cb3515a16d103aa40ba3a3360a34fce