Use spawn_executable when running buck2 commands
Summary:
# Problem
We are current passing arguments as a single string, which is error prone, and we are not escaping files correctly, etc

# This diff
We use `open_port()` with `spawn_executable` to avoid this problem. One reason why this was probably not done before is that we need to do `2> /dev/null`, since `buck2 uquery`  with `--console=none` stupidily suppresses all output.

We workaround this by adding a small trampoline that handles the redirection

Reviewed By: TheGeorge

Differential Revision: D83584213

fbshipit-source-id: a465879b2412757489fd88e93466b0a0317cc82d