2nd attempt to roll out python-based apple installer
Summary:
This diff makes a 2nd attempt in rolling out the Python-based Apple installer.

The first attempt of rolling out this feature (D82853031) caused a sev (S569826) because the apple installer was not compatible with the fat platform and a lot of actions were being redirected to remote mac machines.
As shown here: P1972179763

Previous diffs in this stack fixes this issue by making the new tool compatible with the fat platform.

Reviewed By: chatura-atapattu, benb

Differential Revision: D83660065

fbshipit-source-id: acafb60075c27d49d3b38945a116554d4f4606e8