Delete tbd from headers functionality
Summary: This functionality was used to attempt to generate an interface for a shared library from headers directly. It became too difficult to identify which headers to include, and so the solution was never used. Remove it so it can be replaced with a solution that seeks to do the same, but from object files rather than headers.

Reviewed By: rmaz

Differential Revision: D83865083

fbshipit-source-id: 2b07c801fa350829885a6913f6094a40dd2e9492