Add Kotlinc step execution time into extras
Summary:
### Context
For debugging purposes, it would be helpful to have information about kotlinc time available. Since we have the kotlicd_action table, this information can be easily added to the extras column.

### This diff
Implements the logic of capturing the duration and adding it into the extras column for improved debugging capabilities.

Reviewed By: IanChilds

Differential Revision: D84257977

fbshipit-source-id: 3db5ed2b520a09d423f9e4900d59eecc9a431215