Code Signing: add `[signing-context]` to `apple_bundle()`
Summary:
**Problem**

In certain cases, it's useful to know whether an application will be codesigned and provisioned *without* having to build the application.

**Solution**

Expose a `[signing-context]` subtarget on `apple_bundle()` which provides a JSON output with the necessary code signing and provisioning profile data.

This subtarget does *not* require the bundle to the built, so it's super fast.

Reviewed By: benb

Differential Revision: D84350180

fbshipit-source-id: b0009064a39e8b9d25fe9b55746ccd6f849f5e08