Code Signing: postprocess signing context to include info inline
Summary:
**Problem**

The signing context is much harder to consume because it includes paths to each bundle's signing context. This complicates clients which have to handle all of this logic.

**Solution**

Inline the information from all the signing contexts, in the same way we do for the `[codesign-manifest]`. `[signing-context-tree]` is removed and only `[signing-context]` is left which now includes transitive information.

Reviewed By: benb

Differential Revision: D84497873

fbshipit-source-id: a6264c639e7ed6362213f202705a31a81120addc