Refactor error handling in test_binary
Summary:
# Context
The `test_binary` is in charge of running CT tests. It handles several commands:
- listing tests
- running tests
- listing and running commands

# Problem
Running each command has some error-handling boilerplate to ensure that all logs are flushed, exceptions are logged and an exit code is returned. The logic is copy-pasted on every command, which obfuscate the main logic.

Moreover, we are about to introduce another command, which will make the problem worse

# This diff
Introduce a `run_action_and_halt()` function that contains all the boilerplate, and refactor the commands to use that instead

Reviewed By: robertoaloi

Differential Revision: D85061726

fbshipit-source-id: 7148e3eaac19b09e6ec76a1f7a47b17bc69b2728