Fix transitive prebuilt check
Summary:
`_transitive_has_linkable` only checks if any deps of the target have
linkables, not the target itself. By calling this function on the deps
of the target we're checking for instead of directly on the target, we'd
miss checking whether any of the first-level dependencies of the target
had a linkable. Fix this and rename the function to be clearer.

Reviewed By: mzlee

Differential Revision: D85090784

fbshipit-source-id: 6165b925bfcaf53e2f70c4e589727cfdccb4067a