Use the default isolation directory when building indexes
Summary:
In general you get much better buck caching when you use the default isolation directory, so don't override the isolation directory in the indexer.

This was added in D81300493, but doesn't seem to be necessary any longer. If we hit another case of recursive buck invocations, the outer invocation should set the isolation directory, not the inner invocation which is building all the many swift targets.

Reviewed By: CatherineGasnier

Differential Revision: D85245945

fbshipit-source-id: 9d89363b31da4a4123e5e17f25ae1e71484c5b22