monitor test_binary from the test node
Summary:
When the test_binary is killed by the test runner or the os (e.g. SIGKILL) it might not be able to shutdown the test node properly. In these cases, we have test nodes lingering around and using up resources.

E.g. in a scenario where the test node hangs, and the test_binary is killed (e.g. by timeout) the test node can survive poisoning the sytem.

We install a simple TCP connection from the test node to the test binary. If the test binary exits in any way, the socket closes and the test node terminates itself.

Reviewed By: acw224

Differential Revision: D85196780

fbshipit-source-id: f0f71018e49235d2076513dba4733c72e6d6d450