add subtargets on apple_binary
Summary: Add two subtargets on apple_binary to collect plugin and socket manifests for deps within the main binary. This uses dynamic actions to defer the cost of tset traversal. The set of argsfiles are always added as dependent outputs as there is seemingly no way to extract the DynamicValue providers from the return value.

Reviewed By: d16r

Differential Revision: D85369763

fbshipit-source-id: 3517443584f5c065b01f28b80db238bde7487a4f