Add support for running tests on global epmd
Summary:
# Context
When running tests, we start a dedicated `epmd` on a random port for each suite. This is a feature, as it ensures that suites running in parallel are well isolated.

# Problem
This is however problematic for  running under `edb`. When launching an executable, `edb` currently will ensure `epmd` is started, and expect the debuggee node to establish a dist connection with the `edb` node. If the test node runs on its own `epmd`, then it will not find the `edb` node and the attachment will fail.

# This diff
Add a `--no-epmd` option that will make it so the global epmd (if available) is used, instead of spinning up a new one.

Notice that we prefer to add a new option, instead of associating it to the `--edb-code` option, as it may be possible in the future to launch a debuggee node without requiring distribution

There was already some internal support to use the global epmd via a `test_exec` app env, which makes the change not intrusive

Reviewed By: TheGeorge

Differential Revision: D85249497

fbshipit-source-id: 121e66b46d365d965827ef6b1754b0663b2cf1f8