Use standalone test logging when running under edb
Summary:
# Context
Tests can run stand-alone or under the TPX runner. The way we configure logging is different in each case.

# Problem
When running under edb, we want  logging, stdout, etc to behave as if it were standalone, but we are currently using the TPX settings.

# This diff
Use standalone settings when edb is in use

Reviewed By: acw224

Differential Revision: D85560333

fbshipit-source-id: 4e3dd5af9bcbb95b1f07ea92d18419a0bb3552be