Avoid debug printing from runner node when under edb
Summary:
# Context
When running a test, the test-binary node starts a new test-runner node where the test is actually executed, and passes parameters with what needs to be run, etc.

To make it easier to debug in the case the test-runner node fails to start, it currently emits a `io:format()` with the settings it received etc, and because stdout is forwarded to the test-binray node, this ends up in the test-binary logs.

When running without TPX, we explicitly disable this debug logging, as it otherwise shows up on stdout of every suite execution.

# Problem
We now have a similar problem when running a suite under `edb`, where TPX is not in the middle and we would also get this debug output printed on stdout

# This diff
Use the same environment-variable hack used when running without TPX to disable this debug logging when running under edb

Reviewed By: acw224

Differential Revision: D85560336

fbshipit-source-id: d665e358b9754f4eb57604dd60eec87a34f02f4f